backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
celery
redis
